ef exercises before moving

+++ server/ef-exercises/Entities/Employee.cs
@@ -6,7 +6,7 @@ public class Employee
     public string FirstName { get; set; } = string.Empty;
     public string LastName { get; set; } = string.Empty;
     public string Email { get; set; } = string.Empty;
-    public decimal Salary { get; set; }
+    public double Salary { get; set; }
     public DateTime HireDate { get; set; }
     public int DepartmentId { get; set; }
     public Department Department { get; set; } = null!;
